temp fix for wiremock issue
