+++ domains/transfer/Transfer.API/API.IntegrationTests/Cvr/Api/v2024_01_03/Controllers/CvrControllerTests.cvr_response.json
@@ -8,7 +8,7 @@
         "failed": 0
     },
     "hits": {
-        "total": 1,
+        "total": 2,
         "max_score": 12.518271,
         "hits": [
             {
